Separated rates provider and rates service.
